% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_contigency_table.r
\name{create_dat_two}
\alias{create_dat_two}
\title{create a function to create original data from a table 2x2}
\usage{
create_dat_two(tab, name1, name2)
}
\arguments{
\item{tab}{contigency table 2x2}

\item{name1}{A string that name of first variable into table}

\item{name2}{A string that name of second variable into table}
}
\value{
res a dataframe that has two column
which column 1 is first variable and column 2 is second variable
}
\description{
create a function to create original data from a table 2x2
}
\examples{
\dontrun{
create_dat_two(mytable, "Expose", "Disease")
}
}
