% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_contigency_table.r
\name{draw_ellipse_ci}
\alias{draw_ellipse_ci}
\title{Confidence ellipse for bivariate normal}
\usage{
draw_ellipse_ci(S, xbar, alpha = .05, n = 100)
}
\arguments{
\item{S}{The Covariance Matrix}

\item{xbar}{sample Mean}

\item{alpha}{\eqn{1-\alpha} is level of CI.}

\item{n}{number of observaions}
}
\value{
A plot that draw a ellipse with its diagonals and show sample Mean in center of that.
}
\description{
To draw confidence ellipses for the mean of a two-variable normal distribution,
there are functions in R, but none of them are theoretically accurate. Here, we have tried to draw this
confidence ellipse in accordance with the academic texts.
}
\examples{
\dontrun{
    S <- matrix(c(1, -1, -1, 4), 2, 2)
    xbar <- c(0, 0)
    draw_ellipse_ci(S = S, xbar = xbar, alpha = .05, n = 100)
}

}
\references{
Johnson, R. A., & Wichern, D. W. (1992). Applied multivariate statistical analysis. New Jersey, 405.
}
\author{
Habib Ezatabadi
}
