% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_contigency_table.r
\name{get_contigency_result}
\alias{get_contigency_result}
\title{Function to create a complete table results for contigency table}
\usage{
\code{get_contigency_result(n11, n12, n21, n22,
    varname1 = "Expose", varname2 = "Disease",
    levels_var1 = c("Exposed", "UnExposed"), 
    levels_var2 = c("Disease", "UnDisease"), show_table_results = TRUE)}
}
\arguments{
\item{n11}{The number that shows this is that the first
variable of the table is at its first level and the second
variable of the table is also at its first level}

\item{n12}{The numbers that indicate this, the first
variable of the table is on its
first level and the second variable of
the table is on its second level}

\item{n21}{The numbers that indicate this,
the first variable of the table is on its second
level and the second variable of the table is on its first level}

\item{n22}{The numbers that indicate this, the first variable of
the table is on its second level and the second variable of
the table is also on its second level}

\item{varname1}{name of first variable}

\item{varname2}{name of second variable}

\item{levels_var1}{levels of first variable}

\item{levels_var2}{levels of second variable}

\item{show_table_results}{A logical variable that takes two values,
FALSE and TRUE, when in the TRUE state, is displayed in the output of a
complete table as an HTML page.}
}
\value{
Table_results A list containing 8 output tables in
\href{https://cran.r-project.org/web/packages/kableExtra/vignettes/awesome_table_in_html.html}{kableExtra} format,
showing the outputs for each table.

stat_R_results list of 8 table as dataframe format for show
result of table that generate from contigency table.
}
\description{
Function to create a complete table results for contigency table
}
\examples{
\dontrun{get_contigency_result(475, 461, 7, 61, "Expose", "Disease",levels_var1 = c("Exposed", "UnExposed"), levels_var2 = c("Disease", "UnDisease"), show_table_results = TRUE)}
}
