% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_contigency_table.r
\name{get_dat_from_tab}
\alias{get_dat_from_tab}
\title{This function is designed so that, according to the user's request,
from an Contigency table based on two variables,
a data set with type; Create a matrix or dataframe or list.}
\usage{
\code{get_dat_from_tab(tab, Levels = NULL , idLevel = 0, data_type = "Matrix", 
    varnames = c("Var1", "Var2"))}
}
\arguments{
\item{tab}{contigency table based on Two Variables.}

\item{Levels}{A list with two members, the first member of the variable levels
that is distributed in the rows of the contigency
table and the second Member in its columns,
the default value is NULL. And level two and ... for two variables.}

\item{idLevel}{indicator variable, if the \code{Levels} argument is entered,
this argument must take the value 1, otherwise 0.}

\item{data_type}{According to the user's request, if you want the format of
the output data to be in the form of a matrix,
the value of the \code{"Matrix"} is entered, for
the dataframe, \code{"dataframe"} and for the list entered \code{"list"}.}

\item{varnames}{A vector with two members, which are the names of the
first variable (the variable whose levels are distributed in the
rows of the contigency table) and the second.}
}
\value{
The output is a list with two members, input table (\code{original_table}) and dataset (\code{Data}).
}
\description{
This function is designed so that, according to the user's request,
from an Contigency table based on two variables,
a data set with type; Create a matrix or dataframe or list.
}
\examples{
\dontrun{data(table_2)
    get_dat_from_tab(tab = table_2, data_type = "dataframe")}
}
