% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_contigency_table.r
\name{h_fisher}
\alias{h_fisher}
\title{This function is designed to implement Fisher's algorithm for exact testing in a 2x2 contigency table.
Although the \code{\link[stats:fisher.test]{stats::fisher.test()}} function is a
very fast and good function, this function is also suitable.}
\usage{
\code{h_fisher(tab, alternative = "two-sided")}
}
\arguments{
\item{tab}{contigency table \deqn{2 \times 2}}

\item{alternative}{argumment that can take 3 value ("two-sided", "less", "greater")}
}
\value{
a vector with two element \code{"p-value", "p-table"} that, \code{"p-value" is} \deqn{p_{value}} of test and
\code{p-table} is probablity of original table.
}
\description{
This function is designed to implement Fisher's algorithm for exact testing in a 2x2 contigency table.
Although the \code{\link[stats:fisher.test]{stats::fisher.test()}} function is a
very fast and good function, this function is also suitable.
}
\examples{
\dontrun{tab2 <- matrix(c(1, 9, 11, 3), 2, 2,
    byrow = T)
    h_fisher(tab2, alternative = "two-sided")}
}
