% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_contigency_table.r
\name{homogenity_test_or}
\alias{homogenity_test_or}
\title{this function created for get mantel-haenszel and test homogenty of OR}
\usage{
homogenity_test_or(x, partial_oddsratio_method = "wald",
    confront_var = "age")
}
\arguments{
\item{x}{is array with Atleast 3 dimension}

\item{partial_oddsratio_method}{method The odds ratio estimation method has three state \code{"midp", 
"wald", "exact"}}

\item{confront_var}{confounding variable is A factor variable}
}
\value{
odd_ratio_result result

test_result resut results

tabe_test t table
}
\description{
this function created for get mantel-haenszel and test homogenty of OR
}
\examples{
\dontrun{homogenity_test_or(x, partial_oddsratio_method = "wald", confront_var = "age")}
}
