% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_contigency_table.r
\docType{data}
\name{HypterTension}
\alias{HypterTension}
\alias{HyperTension}
\title{HyperTension Data}
\format{
dataframe
}
\usage{
data(HyperTension)
}
\description{
\if{html}{\out{<div class="sourceCode">}}\preformatted{A dataset with two variables and 200 observations.
}\if{html}{\out{</div>}}

\itemize{
\item value: The variable in which blood pressure
measurement values are stored in two groups of standard treatment and new treatment.
\item Groups: The name of the treatment group
}
}
\keyword{datasets}
