% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_contigency_table.r
\name{ifel}
\alias{ifel}
\title{The base Function of R for applying a condition on a vector at the
same time is in the form that  return is the first value of the vecotr
This function is designed to return a vector by applying a condition on a vector.}
\usage{
\code{ifel(cond, x, y)}
}
\arguments{
\item{cond}{Alogical value, that is TRUE or FALSE}

\item{x}{if \code{cond = TRUE} return \code{x}}

\item{y}{if \code{cond = FALSE} return \code{y}}
}
\description{
The base Function of R for applying a condition on a vector at the
same time is in the form that  return is the first value of the vecotr
This function is designed to return a vector by applying a condition on a vector.
}
\examples{
\dontrun{ifel(TRUE, c(1, 2, 5), c(4, 1, 3))}
}
\seealso{
\code{\link[base:ifelse]{base::ifelse()}}
}
