% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_contigency_table.r
\name{IfElse}
\alias{IfElse}
\title{R's builtin function for applying a condition on a vector at the same time is
in the form that the return is the first value of the vector (\code{\link[base:ifelse]{base::ifelse()}).
This function is designed to return a vector by applying a condition on a vector.}
\usage{
\code{IfElse(cond, x, y)}
}
\arguments{
\item{cond}{A logical value, that is TRUE or FALSE}

\item{x}{if \code{cond == TRUE} return \code{x}}

\item{y}{if \code{cond == FALSE} return \code{y}}
}
\value{
\code{x} or \code{y}
}
\description{
R's builtin function for applying a condition on a vector at the same time is
in the form that the return is the first value of the vector (\code{\link[base:ifelse]{base::ifelse()}).
This function is designed to return a vector by applying a condition on a vector.
}
\examples{
\dontrun{IfElse(TRUE, c(1, 2, 5), c(4, 1, 3))}
}
