% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_contigency_table.r
\name{Inferiority_superiority_test_pa}
\alias{Inferiority_superiority_test_pa}
\title{Non-Inferiority and superiority Test in cilinical Trials, for compare two Treatment}
\usage{
Inferiority_superiority_test_pa(
    dataType  = "binary",  Dat, alpha = .05, 
    Method_estimate_for_binary_data = "fm", 
    margin, reff = 1, better = "right", Test_Method = "N", 
    Name_groups = c(group1 = "standard", group2 = "new"), 
    data_list = NA
    )
}
\arguments{
\item{dataType}{It can take two values, \code{c("binary", "continuous")}, to tell the
function whether our data is to compare rates or continuous values.}

\item{Dat}{is our data set, the thing about it is that the data must have two columns,
the first column is related to values and the second column is related to factors,
that is, there must be a specific factor in the second column, which determines that observation
It is related to which treatment, the next point is that the agents can only be of two types.
And also for binary data, the value column should have only  \code{TRUE} (success) and \code{FALSE} (failure) values for each observation.
if Dat take NA value, that means our data is in data_list list.}

\item{alpha}{\eqn{0 < \alpha < 1}, level of test.}

\item{Method_estimate_for_binary_data}{It can take 3 values, \code{c("fm", "ha", "wald")},
it should be noted that \code{"fm"} abbreviated of \code{"Farrington-Manning"} Method,
\code{"ha"} abbreviated of \code{"Hauck-Anderson"} method, and \code{"wald"} implements the famous \code{"wald"} method for binary data.}

\item{margin}{\eqn{\delta,~H_0: p_1 - p_2 \leq \delta ~Vs~H_1: p_1 - p_2 > \delta~} Null-value for more information about margin in Non-Inferiority

and Superiority test go to \code{details.}}

\item{reff}{It takes two values \code{reff = 1, reff = 2}, 1 means that we want to subtract the average of the second community from the first and 2 means that
we want to subtract the difference of the first community from the second. if \code{reff = 1} then \eqn{H_0: \mu_1 - \mu_2 \leq \delta}
\code{reff = 2} then \eqn{H_0: \mu_2 - \mu_1 \leq \delta}.}

\item{better}{It takes two values \code{c("left", "right")}. Maybe for certain tests,
a lower value indicates a better treatment, in which case we should give this argument the \code{"left"} value,
and if a higher value indicates a better treatment, by default, this argument contains the \code{"right"} value.}

\item{Test_Method}{getting two value \code{c("N", "S")} N infer to \code{"Non-Inferiority Test"} and S infer to \code{"Superiority Test"}.}

\item{Name_groups}{A Vector with Two element, name of group 1 and name of group 2.
It is necessary that the values of this argument be quantified. By default,
the values of this argument are \code{c("standard", "new")} standard for group 1 and new for group two.}

\item{data_list}{This is A list; if \code{Dat == NA}, this argument can handle data for both binary and continuous data,
For binary data, this list takes 4 values, N11, N12, N21 and N22, of course, the order of input must be exactly the same,
where N11 refers to the number of successes related to the first group and N12 refers to the number of failures in the first group.
N21 refers to the number of successes of the second group and N22 refers to the number of failures of the second group,
but if the data is continuous, this list has two members, the first of which is X or the values related to the first group and the second Y,
which contains the values related to the second group. has it.
in default; data_list = NA and Dat argument take data.}
}
\value{
A List of Three Components
\tabular{ll}{
\code{TestResult}:\tab "htest" class that contains result of test\cr
\code{plotResult}:\tab Plot of Test \cr
\code{TestTable}:\tab A Html table for show results \cr
}

TestResult that from class "htest" contains below members:\
\tabular{ll}{
\code{statistic}:\tab the value of the Z-statistic\cr
\code{parameter}:\tab delta, rate difference (group 1 - group 2) under the null hypothesis\cr
\code{p.value}:\tab the p-value for the Farrington-Manning test\cr
\code{null.value}:\tab rate difference (group 1 - group 2) under the null\cr
\code{alternative}:\tab a character string indicating the alternative hypothesis\cr
\code{method}:\tab a character string indicating the exact method employed\cr
\code{data.name}:\tab a character string giving the names of the data used\cr
\code{estimate}:\tab the estimated rate difference (maximum likelihood)\cr
\code{conf.int}:\tab a confidence interval for the rate difference\cr
\code{sample.size}:\tab the total sample size used for the test\cr
}
}
\description{
In order to implement non-inferiority tests and superiority tests,
in randomized clinical trials, when we want to compare two types of treatment or two types of drugs or a drug with a placebo,
there are many softwares, there are various functions in R, some From the functions used in R,
they do not have optimal outputs and even sometimes, their outputs are not very correct. In this function,
we have tried to have integrated outputs with a completely simple table, as well as adding a graph,
two non-derogatory tests. and superiority in studies designed for parallel groups.
}
\details{
for Non-Inferiority and Superiority Test for rates and continuous values like
Hypertension, We set a null value based on previous clinical studies or articles and studies already done, then:
\describe{
\item{Non-Inferiority}{If a treatment is better, i.e. that treatment has a larger mean in
recorded observations or a higher rates, our test is defined as:
\deqn{\text{for rates}: \quad \begin{cases}H_0: & p_1 - p_2 \leq \delta \\
H_1: & p_1 - p_2 > \delta, \end{cases} \quad \delta < 0.
}
\deqn{
\text{or for continuous values}: \quad \begin{cases}H_0: & \text{Treat}_1 - \text{Treat}_2 \leq \delta \\
H_1: & \text{Treat}_1 - \text{Treat}_2 > \delta \end{cases} \quad \delta < 0.
}
If a treatment is better, i.e. that treatment has a smaller mean in the
recorded observations or a smaller rates, our test is defined as:
\deqn{\text{for rates}: \quad \begin{cases}H_0: & p_1 - p_2 \geq \delta \\
H_1: & p_1 - p_2 < \delta, \end{cases} \quad \delta > 0. }
\deqn{
\text{or for continuous values}: \quad \begin{cases}H_0: & \text{Treat}_1 - \text{Treat}_2 \geq \delta \\
H_1: & \text{Treat}_1 - \text{Treat}_2 < \delta \end{cases} \quad \delta > 0.
}
}
\item{Superiority}{If a treatment is better, i.e. that treatment has a larger mean in
recorded observations or a higher rates, our test is defined as:
\deqn{\text{for rates}: \quad \begin{cases}H_0: & p_1 - p_2 \leq \delta \\
H_1: & p_1 - p_2 > \delta, \end{cases} \quad \delta > 0. }
\deqn{
\text{or for continuous values}: \quad \begin{cases}H_0: & \text{Treat}_1 - \text{Treat}_2 \leq \delta \\
H_1: & \text{Treat}_1 - \text{Treat}_2 > \delta \end{cases} \quad \delta > 0.
}
If a treatment is better, i.e. that treatment has a smaller mean in the
recorded observations or a smaller rates, our test is defined as:
\deqn{\text{for rates}: \quad \begin{cases}H_0: & p_1 - p_2 \geq \delta \\
H_1: & p_1 - p_2 < \delta, \end{cases} \quad \delta < 0. }
\deqn{
\text{or for continuous values}: \quad \begin{cases}H_0: & \text{Treat}_1 - \text{Treat}_2 \geq \delta \\
H_1: & \text{Treat}_1 - \text{Treat}_2 < \delta \end{cases} \quad \delta < 0.
}
}
\item{Rates}{
for rates we add three method in this code;
\code{"wald"}, \code{"Farrington-Manning"} and \code{"Hauck-Anderson"} for get
more information about this methods
you can go to this \href{https://1drv.ms/b/s!AjGgWTomcc6MgRw7N9ZTMJjKAEvO?e=SMFLgT}{link.}
}
\item{Continuous Values}{
for continuous values we implement a \code{T Test}:
\deqn{
\bar{x}_1 = \frac{1}{n_1}\sum_{i = 1}^{n_{1}} x_{i_{1}}, \quad \bar{x}_2 = \frac{1}{n_{2}}\sum_{i = 1}^{n_2} x_{i_{2}}
}
\deqn{
\bar{x}_1 \sim \mathcal{N}(\mu_1, \frac{\sigma_1}{n_1}), \quad \bar{x}_2 \sim \mathcal{N}(\mu_2, \frac{\sigma_2}{n_2}),
}
\deqn{
\text{we want to test}:\quad \begin{cases}H_0: & \mu_1 - \mu_2 \leq \delta \\
H_1: & \mu_1 - \mu_2 > \delta\end {cases},
}
\deqn{
\text{if}~~ \sigma_1 = \sigma_2 \implies S^2_{pooled} = \frac{(n_1 -1)S^2_1 + (n_2 -1)S^2_2}{n_1 + n_2 -2},}
\deqn{
S^2_1 = \frac{1}{n_1 - 1}\sum_{i = 1}^{n_1}(x_{i_{1}} - \bar{x}_1)^2,}
\deqn{
S^2_2 = \frac{1}{n_2 - 1}\sum_{i = 1}^{n_2}(x_{i_{2}} - \bar{x}_2)^2, \implies
}
\deqn{
\text{Test Statistics}: \frac{\bar{x}_1 - \bar{x}_2 - \delta}{s_{pooled}\sqrt{\frac{1}{n_1}+ \frac{1}{n_2}}} \sim \underset{\text{if}~H_0 ~\text{is TRUE}}{\sim} T_{(n_1 + n_2 - 2)}
}
\deqn{
\text{if}~ \sigma_1 \ne \sigma_2 \implies
}
We have to use Welch-Satterthwaite, so for more information about this statistics, go to this \href{https://en.wikipedia.org/wiki/Welch\%27s_t-test}{link.}
}
}
}
\examples{
\dontrun{
    data(HyperTension)
    Inferiority_superiority_test_pa(dataType = "continuous", Dat = HyperTension, alpha = 0.05,
    margin = 5, reff = 2, better = "right", Test_Method = "N", 
    Name_groups  = c(group1 = "standard", group2 = "new"), 
    data_list = NA)
}

}
\author{
Habib Ezatabadi
}
