% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_contigency_table.r
\name{lambda_coef_contigency}
\alias{lambda_coef_contigency}
\title{define function for get Lambda coefficients}
\usage{
lambda_coef_contigency(n11, n12, n21, n22, 
    varname1 = "Expose", varname2 = "Diseasee", levels_var1 = c("Exposed", 
    "UnExposed"), levels_var2 = c("Disease", "UnDisease"))
}
\arguments{
\item{n11}{see also \code{\link{get_contigency_result}}}

\item{n12}{see also \code{\link{get_contigency_result}}}

\item{n21}{see also \code{\link{get_contigency_result}}}

\item{n22}{see also \code{\link{get_contigency_result}}}

\item{varname1}{see also \code{\link{get_contigency_result}}}

\item{varname2}{see also \code{\link{get_contigency_result}}}

\item{levels_var1}{see also \code{\link{get_contigency_result}}}

\item{levels_var2}{see also \code{\link{get_contigency_result}}}
}
\value{
table of lambda result, for more detail of what is lambda
}
\description{
define function for get Lambda coefficients
}
\examples{
\dontrun{lambda_coef_contigency(475, 461, 7, 61, "Expose", "Disease", 
    levels_var1 = c("Exposed", "UnExposed"), 
    levels_var2 = c("Disease", "UnDisease"))}
}
