% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_contigency_table.r
\docType{data}
\name{rct_binary_data_1}
\alias{rct_binary_data_1}
\title{rct_binary_data_1 Data}
\format{
dataframe
}
\usage{
data(rct_binary_data_1)
}
\description{
\if{html}{\out{<div class="sourceCode">}}\preformatted{A dataset with two variables and 114 observations.
}\if{html}{\out{</div>}}

\itemize{
\item values: Recovery or non-recovery for the treatment group.
\item group: The name of the treatment group
}
}
\keyword{datasets}
