% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_contigency_table.r
\docType{data}
\name{rct_binary_data_2}
\alias{rct_binary_data_2}
\title{rct_binary_data_2 Data}
\format{
dataframe
}
\usage{
data(rct_binary_data_2)
}
\description{
\if{html}{\out{<div class="sourceCode">}}\preformatted{A Table with 3 variables.
}\if{html}{\out{</div>}}

\itemize{
\item drug: A type of drug
\item alive: alive = 1, death = 0
\item count number of patients that alive or death.
}
}
\keyword{datasets}
