% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_contigency_table.r
\docType{data}
\name{rct_continuous_data_3}
\alias{rct_continuous_data_3}
\title{rct_continuous_data_3 Data}
\format{
dataframe
}
\usage{
data(rct_continuous_data_3)
}
\description{
\if{html}{\out{<div class="sourceCode">}}\preformatted{A dataset with two variables and 22 observations.
}\if{html}{\out{</div>}}

\itemize{
\item values: Measured criteria for each patient's recovery
\item treat: The Type of Treatment
}
}
\keyword{datasets}
