% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_contigency_table.r
\name{stats_round}
\alias{stats_round}
\title{Rounding vectors that have multi-type values (characters and numbers)}
\usage{
stats_round(x, ndigit = 4)
}
\arguments{
\item{x}{A vector, with numeric and character elements}

\item{ndigit}{The number of decimal digits we want to round the numbers inside the vector}
}
\value{
A vector of the same size but with numbers rounded to an arbitrary number of decimal places.
}
\description{
To round a vector that has both numeric values and character values,
we know that if a vector contains characters, all the values have character
format, but sometimes we need to round the numeric values to several decimal
places when displaying the vector. This function is implemented to round
numerical values in vectors that contain characters.
}
\examples{
\dontrun{stats_round(x = c("a", 2.342341, "stats9", 3.324234235), ndigit = 2)}
}
\seealso{
\code{\link[base:round]{base::round()}}
}
\author{
Habib Ezatabadi
}
