% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_contigency_table.r
\docType{data}
\name{table_1}
\alias{table_1}
\title{table_1 contigency table with 3 variables}
\format{
contigency table with 3 variables
}
\usage{
data(table_1)
}
\description{
A dataset containing a contigency table with 3 variable
The variables are as follows:
}
\details{
\itemize{
\item exposure. The variable that shows how many were exposed, which is a binary variable with two levels of exposure (1) or no exposure (0).
\item Group. A binary variable that is leveled at the level of the treated group (1) and the control group (0).
\item age. A categorical variable, which is divided into three levels: 1, 2, and 3.
}
}
\keyword{datasets}
