% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_contigency_table.r
\docType{data}
\name{table_2}
\alias{table_2}
\title{table_2 contigency table with 2 variables}
\format{
contigency table with 2 variables
}
\usage{
data(table_2)
}
\description{
A contigency table based on the number of case-control study for ovarian cancer patients
and its association with contraceptive use and duration of use.
}
\details{
\itemize{
\item Disease: The variable that shows how many were Disease (case) on Not Disease (control),
which is a binary variable with two levels of Disease (case) or Not Disease (control).
\item OC Duration time: How long the person in question has been using contraceptives.
which has 4 levels, no use (\code{None}), between 0 and 5 years of use (\code{0-5}),
between 5 and 10 years of use (\code{50-10} and more than 10 years of use (\code{>10}).
}
}
\keyword{datasets}
