% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_contigency_table.r
\name{list_to_dataframe}
\alias{list_to_dataframe}
\title{this function created for convert a list to dataframe,
List members must be vectors with equal number of members.}
\usage{
\code{list_to_dataframe(mylist)}
}
\arguments{
\item{list}{a list; List members must be vectors with equal number of members}
}
\value{
a dataframe, A dataframe whose columns are members of the input list.
}
\description{
this function created for convert a list to dataframe,
List members must be vectors with equal number of members.
}
\examples{
\dontrun{list_to_dataframe(mylist)}
}
