% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_contigency_table.r
\name{Table_Test_Result}
\alias{Table_Test_Result}
\title{This function has been prepared for the purpose of performing three valid tests to check the connection
or non-connection of the columns and rows of a contigency table and to output the
test statistics as well as the expected values of the table and to check whether
the exact test should also be performed or not. bring.}
\usage{
\code{Table_Test_Result(tab, Levels, idLevel = 0)}
}
\arguments{
\item{tab}{contigency table with two variable, that any variable have I (I >= 2) levels.}

\item{Levels}{see \code{\link{get_dat_from_tab}}}

\item{idLevel}{see \code{\link{get_dat_from_tab}}}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{\\itemize\{
     \\item ExpEcted_Vals: table of expected values of a contigency table (tab)
     \\item test_result: table of test results
     \\item Total_results: table of Total results (expected values, test resutls and input table)
     \\item table_results: html table for total results
}\if{html}{\out{</div>}}

}
}
\description{
This function has been prepared for the purpose of performing three valid tests to check the connection
or non-connection of the columns and rows of a contigency table and to output the
test statistics as well as the expected values of the table and to check whether
the exact test should also be performed or not. bring.
}
\details{
for calculate test statistics values, we use this formulas:
\deqn{
        \begin{align}
        & \text{Contigency Table} = \left[\begin{array}{c|c|c|c}n_{(1, ~1)} & n_{(1,~2)} & \cdots & n_{(1,~J)}\\
        n_{(2, ~1)} & n_{(2, ~2)} & \cdots & n_{(2, ~J)} \\
        \vdots & \ddots & \ddots & \vdots \\
        n_{(I, ~1)} & n_{(2, ~2)} & \cdots & n_{(I, ~J)}\end{array}\right]\\
        & \Lambda = \frac{\prod_i \prod_j(n_{i+}\times n_{+j})^{n_{ij}}}{n \prod_i\prod_j n_{ij}^{n_{ij}}}\\
        & G^2 = -2\log(\Lambda) = 2\sum_i\sum_j n_{ij}\log\left(\frac{n_{ij}}{\hat{\mu}_{ij}}\right)\\
        & \hat{\mu}_{ij} = \frac{n_{i+} \times n_{+j}}{n}\\
        & \underset{\text{If}~ H_0 ~ \text{is TRUE}}{G^2} \approx \chi^2_{(I-1)\times (J-1)}\\
        & \chi^2_{\text{pearson}} = \sum_{i = 1}^I\sum_{j = 1}^J\frac{(n_{(i, ~j)}-\hat{\lambda}_{(i, ~j))})^2}{\hat{\lambda}_{(i, ~j))}}\\
        & \underset{\text{If}~ H_0 ~ \text{is TRUE}}{\chi^2_{(\text{pearson})}} \approx \chi^2_{(I-1)\times (J-1)}\\
        & \text{Trend Test Statistics} = M^2 = r^2 \times (n-1)\\
        & \underset{\text{If} ~H_0 ~ \text{Is TRUE}}{M^2} \approx \chi^2_{(1)} \\
        & n = \sum_{i =1}^I\sum_{j=1}^J n_{(i, ~j)}, \\
        & r = \text{Corr}(X_1, ~X_2), \quad X_1, ~ X_2 ~\text{Are two variables of contigency table}
        \end{align}
}
}
\examples{
\dontrun{data(table_2)
    Table_Test_Result(tab = table_2)}
}
